add user can change page of user

+++ app/admin/page.rb
@@ -1,5 +1,5 @@
 ActiveAdmin.register Page do
-menu priority: 2,:if => proc{ current_admin_user.is_admin? },label: "TRANG"
+menu priority: 2,label: "TRANG"
 #menu false
 # See permitted parameters documentation:
 # https://github.com/activeadmin/activeadmin/blob/master/docs/2-resource-customization.md#setting-up-strong-parameters
@@ -16,6 +16,16 @@ menu priority: 2,:if => proc{ current_admin_user.is_admin? },label: "TRANG"
 controller do
     def permitted_params
       params.permit page: [:name,:description,:slogan,:menu_id,:url]
+    end
+      def scoped_collection
+      # some stuffs
+      if current_admin_user.is_admin==1
+        super.all
+      else
+
+        super.where(:id=>current_admin_user.page_id)
+      end
+     
     end
     before_filter { @page_title = "Thêm trang" }
     def edit
@@ -40,7 +50,9 @@ controller do
       links = ''.html_safe
       links += link_to 'Hiển thị', resource_path(resource), :class => "member_link view_link"
       links += link_to 'Sửa', edit_resource_path(resource), :class => "member_link edit_link"
-      links += link_to 'Xóa', resource_path(resource), :method => :delete, :confirm => I18n.t('active_admin.delete_confirmation'), :class => "member_link delete_link"
+      if current_admin_user.is_admin==1
+         links += link_to 'Xóa', resource_path(resource), :method => :delete, :confirm => I18n.t('active_admin.delete_confirmation'), :class => "member_link delete_link"
+    end
       links
     end
     #actions
@@ -50,7 +62,11 @@ controller do
     f.inputs do
       f.input :name, :label => "Tên trang"
       f.input :slogan, :label => "Slogan"
-      f.input :menu, :label => "Menu", :include_blank => false
+      if current_admin_user.is_admin==1
+        f.input :menu, :label => "Menu", :include_blank => false
+    else
+        f.input :menu, :label => "Menu", :include_blank => false,:input_html=>{:disabled=>true}
+    end
       f.label :description,"Mô tả :"
       f.cktext_area :description, :class => 'ckeditor', :label => "Mô tả"
       end
@@ -58,11 +74,12 @@ controller do
   end
 
   config.clear_action_items!
-
+  
   action_item :only => :index do
+    if current_admin_user.is_admin==1
       link_to "Thêm trang" , "/admin/pages/new" 
+    end
   end
-
   show title:"Thông tin chi tiết" do |s|
     panel "Chi tiết trang" do
       attributes_table_for page do
